feat: Add type `Trade` to better verify a trade has occurred.
This commit introduces the type `Trade` that will contain data such as a `tradeId`, `buyOrderId`, and more.
The purpose of this type is to better verify that a trade has occurred, especially in our unit testing where print statements cannot be shown in the terminal. Instead, I'll check if `trades`, a vector of type `Trade`, is not empty.