feat: Implement the ability to cancel orders
This commit will features the ability to cancel orders through the `OrderBook` public member function `cancelOrders`.
It takes an `order`'s `orderId` and maps it into `OrderMap_`, an unordered map.
Each orderId corressponds to an iterator either in `asks_` or `bids_`

Also implemented tests to verify that the `cancelOrders` function indeed works.